test: added AdminChecker permission test
